# swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "The ETSI MEC ISG MEC013 Location API described using OpenAPI. The\
    \ API is based on the Open Mobile Alliance's specification RESTful Network API\
    \ for Zonal Presence"
  version: "1.1.1"
  title: "Location API"
  license:
    name: "ETSI Forge copyright notice"
    url: "https://forge.etsi.org/etsi-forge-copyright-notice.txt"
host: "127.0.0.1:8081"
basePath: "/exampleAPI/location/v1/"
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /zones/{zoneId}:
    get:
      description: "Used to get the status of a zone."
      operationId: "zones_zone_id_get"
      produces:
      - "application/json"
      parameters:
      - name: "zoneId"
        in: "path"
        description: "Zone ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "Successful response to a query regarding the status of a zone"
          examples:
            application/json:
              zoneInfo:
                zoneId: "zone01"
                numberOfAccessPoints: "3"
                numberOfUnserviceableAccessPoints: "1"
                numberOfUsers: "10"
                resourceURL: "http://example.com/exampleAPI/location/v1/zones/zone01"
          schema:
            properties:
              zoneInfo:
                $ref: "#/definitions/ZoneInfo"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /zones/{zoneId}/accessPoints:
    get:
      description: "Access point status can be retrieved for sets of access points\
        \ matching attribute in the request."
      operationId: "zones_zone_id_access_points_get"
      produces:
      - "application/json"
      parameters:
      - name: "zoneId"
        in: "path"
        description: "Zone ID"
        required: true
        type: "string"
      - name: "interestRealm"
        in: "query"
        description: "Interest realm of access point (e.g. geographical area, a type\
          \ of industry etc.)."
        required: false
        type: "string"
      responses:
        "200":
          description: "Successful response to a query a named set of access point\
            \ status request"
          examples:
            application/json:
              accessPointList:
                zoneId: "zone01"
                accessPoint:
                - accessPointId: "001010000000000000000000000000001"
                  locationInfo:
                    latitude: "90.123"
                    longitude: "80.123"
                    altitude: "10.0"
                    accuracy: "0"
                  connectionType: "Macro"
                  operationStatus: "Serviceable"
                  numberOfUsers: "5"
                  interestRealm: "LA"
                  resourceURL: "http://example.com/exampleAPI/location/v1/zones/zone01/accessPoints/ap001"
                - accessPointId: "001010000000000000000000000000010"
                  locationInfo:
                    latitude: "91.123"
                    longitude: "81.123"
                    altitude: "12.0"
                    accuracy: "1"
                  connectionType: "Macro"
                  operationStatus: "Unserviceable"
                  numberOfUsers: "0"
                  interestRealm: "DC"
                  resourceURL: "http://example.com/exampleAPI/location/v1/zones/zone01/accessPoints/ap002"
                - accessPointId: "001010000000000000000000000000011"
                  locationInfo:
                    latitude: "93.123"
                    longitude: "83.123"
                    altitude: "16.0"
                    accuracy: "3"
                  connectionType: "Macro"
                  operationStatus: "Serviceable"
                  numberOfUsers: "5"
                  interestRealm: "NJ"
                  resourceURL: "http://example.com/exampleAPI/location/v1/zones/zone01/accessPoints/ap003"
                resourceURL: "http://example.com/exampleAPI/location/v1/zones/zone01/accessPoints"
          schema:
            properties:
              accessPointList:
                $ref: "#/definitions/AccessPointList"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /zones/{zoneId}/accessPoints/{accessPointId}:
    get:
      description: "Access point status can be retrieved for sets of access points\
        \ matching attribute in the request."
      operationId: "zones_zone_id_access_points_access_point_id_get"
      produces:
      - "application/json"
      parameters:
      - name: "zoneId"
        in: "path"
        description: "Zone ID"
        required: true
        type: "string"
      - name: "accessPointId"
        in: "path"
        description: "Access Point ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "Successful response to a query a named set of access point\
            \ status request"
          examples:
            application/json:
              accessPointInfo:
                accessPointId: "001010000000000000000000000000001"
                locationInfo:
                  latitude: "90.123"
                  longitude: "80.123"
                  altitude: "10.0"
                  accuracy: "0"
                connectionType: "Macro"
                operationStatus: "Serviceable"
                numberOfUsers: "5"
                interestRealm: "LA"
                resourceURL: "http://example.com/exampleAPI/location/v1/zones/zone001/accessPoints/ap001"
          schema:
            properties:
              accessPointInfo:
                $ref: "#/definitions/AccessPointInfo"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /users:
    get:
      description: "Users currently using a zone may be retrieved for sets of access\
        \ points matching attribute in the request"
      operationId: "users_get"
      produces:
      - "application/json"
      parameters:
      - name: "zoneId"
        in: "query"
        description: "Zone ID"
        required: true
        type: "string"
      - name: "accessPointId"
        in: "query"
        description: "Identifier of access point, reference \"definitions\" for string\
          \ format"
        required: false
        type: "string"
      responses:
        "200":
          description: "Successful response to a query users within a zone request"
          examples:
            application/json:
              userList:
                user:
                - address: "acr:192.0.2.1"
                  accessPointId: "001010000000000000000000000000001"
                  zoneId: "zone01"
                  resourceURL: "http://example.com/exampleAPI/location/v1/users/acr%3A192.0.2.1"
                - address: "acr:192.0.2.2"
                  accessPointId: "001010000000000000000000000000001"
                  zoneId: "zone01"
                  resourceURL: "http://example.com/exampleAPI/location/v1/users/acr%3A192.0.2.2"
                - address: "acr:192.0.2.3"
                  accessPointId: "001010000000000000000000000000010"
                  zoneId: "zone01"
                  resourceURL: "http://example.com/exampleAPI/location/v1/users/acr%3A192.0.2.3"
                - address: "acr:192.0.2.4"
                  accessPointId: "001010000000000000000000000000001"
                  zoneId: "zone02"
                  resourceURL: "http://example.com/exampleAPI/location/v1/users/acr%3A192.0.2.4"
                - address: "acr:192.0.2.5"
                  accessPointId: "001010000000000000000000000000010"
                  zoneId: "zone02"
                  resourceURL: "http://example.com/exampleAPI/location/v1/users/acr%3A192.0.2.5"
                resourceURL: "http://example.com/exampleAPI/location/v1/users"
          schema:
            properties:
              userList:
                $ref: "#/definitions/UserList"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /users/{userId}:
    get:
      description: "Users currently using a zone may be retrieved for sets of access\
        \ points matching attribute in the request"
      operationId: "users_user_id_get"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "User ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "Successful response to a query users within a zone request"
          examples:
            application/json:
              userInfo:
                address: "acr:192.0.2.1"
                accessPointId: "001010000000000000000000000000001"
                zoneId: "zone01"
                resourceURL: "http://example.com/exampleAPI/location/v1/users/acr%3A192.0.2.1"
                locationInfo:
                  latitude: "90.123"
                  longitude: "80.123"
                  altitude: "10.0"
                  accuracy: "0"
                contextLocationInfo: "GroundFloor"
          schema:
            properties:
              userInfo:
                $ref: "#/definitions/UserInfo"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /subscriptions/zonalTraffic:
    get:
      description: "This operation is used for retrieving all active subscriptions\
        \ to zonal traffic change notifications."
      operationId: "subscriptions_zonal_traffic_get"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Response to retrieve zonal traffic subscriptions"
          examples:
            application/json:
              notificationSubscriptionList:
                zonalTrafficSubscription:
                - clientCorrelator: "0123"
                  resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zonalTraffic/subscription123"
                  callbackReference:
                    notifyURL: "http://clientApp.example.com/location_notifications/123456"
                  zoneId: "zone01"
                  interestRealm: "LA"
                  userEventCriteria: "Transferring"
                - clientCorrelator: "0124"
                  resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zonalTraffic/subscription124"
                  callbackReference:
                    notifyURL: "http://clientApp.example.com/location_notifications/123457"
                  zoneId: "zone02"
                  interestRealm: "LA"
                  userEventCriteria: "Transferring"
                resourceURL: "http://example.com/exampleAPI/location/v1/zonalTraffic"
          schema:
            $ref: "#/definitions/inline_response_200"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
    post:
      description: "This operation is used for creating a new subscription to zonal\
        \ traffic change notification."
      operationId: "subscriptions_zonal_traffic_post"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "zonalTrafficSubscription"
        description: "Zonal Traffic Subscription"
        required: true
        schema:
          $ref: "#/definitions/ZonalTrafficSubscription"
      responses:
        "201":
          description: "Response to create new zonal traffic subscription"
          examples:
            application/json:
              zonalTrafficSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zonalTraffic/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                zoneId: "zone01"
                interestRealm: "LA"
                userEventCriteria: "Transferring"
          schema:
            properties:
              zonalTrafficSubscription:
                $ref: "#/definitions/ZonalTrafficSubscription"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /subscriptions/zonalTraffic/{subscriptionId}:
    get:
      description: "This operation is used for updating an individual subscription\
        \ to zonal traffic change notification."
      operationId: "subscriptions_zonal_traffic_subscription_id_get"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "Response to retrieve individual zonal traffic subscription"
          examples:
            application/json:
              zonalTrafficSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zonalTraffic/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                zoneId: "zone01"
                interestRealm: "LA"
                userEventCriteria: "Transferring"
          schema:
            properties:
              zonalTrafficSubscription:
                $ref: "#/definitions/ZonalTrafficSubscription"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
    put:
      description: "This operation is used for updating an individual subscription\
        \ to zonal traffic change notification."
      operationId: "subscriptions_zonal_traffic_subscription_id_put"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      - in: "body"
        name: "zonalTrafficSubscription"
        description: "Zonal Traffic Subscription"
        required: true
        schema:
          $ref: "#/definitions/ZonalTrafficSubscription"
      responses:
        "200":
          description: "Response to update individual zonal traffic subscription"
          examples:
            application/json:
              zonalTrafficSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zonalTraffic/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                zoneId: "zone01"
                interestRealm: "LA"
                userEventCriteria: "Transferring"
          schema:
            properties:
              zonalTrafficSubscription:
                $ref: "#/definitions/ZonalTrafficSubscription"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
    delete:
      description: "This operation is used for cancelling a subscription and stopping\
        \ corresponding notifications."
      operationId: "subscriptions_zonal_traffic_subscription_id_delete"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      responses:
        "204":
          description: "No content"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /subscriptions/userTracking:
    get:
      description: "This operation is used for retrieving all active subscriptions\
        \ to user tracking change notifications."
      operationId: "subscriptions_user_tracking_get"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Response to retrieve user tracking subscriptions"
          examples:
            application/json:
              notificationSubscriptionList:
                userTrackingSubscription:
                - clientCorrelator: "0123"
                  resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
                  callbackReference:
                    notifyURL: "http://clientApp.example.com/location_notifications/123456"
                  address: "acr:192.0.2.1"
                  userEventCriteria: "Transferring"
                - clientCorrelator: "0124"
                  resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription124"
                  callbackReference:
                    notifyURL: "http://clientApp.example.com/location_notifications/123456"
                  address: "acr:192.0.2.2"
                  userEventCriteria: "Transferring"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking"
          schema:
            $ref: "#/definitions/inline_response_200_1"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
    post:
      description: "This operation is used for creating a new subscription to user\
        \ tracking change notification"
      operationId: "subscriptions_user_tracking_post"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "userTrackingSubscription"
        description: "User Tracking Subscription"
        required: true
        schema:
          $ref: "#/definitions/UserTrackingSubscription"
      responses:
        "201":
          description: "Response to create new user tracking subscription"
          examples:
            application/json:
              userTrackingSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                address: "acr:192.0.2.1"
                userEventCriteria: "Transferring"
          schema:
            properties:
              userTrackingSubscription:
                $ref: "#/definitions/UserTrackingSubscription"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /subscriptions/userTracking/{subscriptionId}:
    get:
      description: "This operation is used for retrieving an individual subscription\
        \ to user tracking change notification."
      operationId: "subscriptions_user_tracking_subscription_id_get"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      responses:
        "200":
          description: "Response to retrieve individual user tracking subscription"
          examples:
            application/json:
              userTrackingSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                address: "acr:192.0.2.1"
                userEventCriteria: "Transferring"
          schema:
            properties:
              userTrackingSubscription:
                $ref: "#/definitions/UserTrackingSubscription"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
    put:
      description: "This operation is used for updating an individual subscription\
        \ to user tracking change notification."
      operationId: "subscriptions_user_tracking_subscription_id_put"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      - in: "body"
        name: "userTrackingSubscription"
        description: "User Tracking Subscription"
        required: true
        schema:
          $ref: "#/definitions/UserTrackingSubscription"
      responses:
        "200":
          description: "Response to update individual user tracking subscription"
          examples:
            application/json:
              userTrackingSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/userTracking/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                address: "acr:192.0.2.1"
                userEventCriteria: "Transferring"
          schema:
            properties:
              userTrackingSubscription:
                $ref: "#/definitions/UserTrackingSubscription"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
    delete:
      description: "This operation is used for retrieving an individual subscription\
        \ to user tracking change notification."
      operationId: "subscriptions_user_tracking_subscription_id_delete"
      produces:
      - "application/json"
      parameters:
      - name: "subscriptionId"
        in: "path"
        description: "Subscription ID"
        required: true
        type: "string"
      responses:
        "204":
          description: "No Content"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
  /subscriptions/zonalStatus:
    get:
      description: "This operation is used for creating a new subscription to zone\
        \ status change notification."
      operationId: "subscriptions_zonal_status_get"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Response to retrieve zone status subscriptions"
          examples:
            application/json:
              notificationSubscriptionList:
                zoneStatusSubscription:
                - clientCorrelator: "0123"
                  resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zoneStatus/subscription123"
                  callbackReference:
                    notifyURL: "http://clientApp.example.com/location_notifications/123456"
                  zoneId: "zone01"
                  numberOfUsersZoneThreshold: "500"
                  operationStatus: "Serviceable"
                - clientCorrelator: "0124"
                  resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zoneStatus/subscription124"
                  callbackReference:
                    notifyURL: "http://clientApp.example.com/location_notifications/123457"
                  zoneId: "zone02"
                  numberOfUsersAPThreshold: "50"
                  operationStatus: "Serviceable"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zoneStatus"
          schema:
            $ref: "#/definitions/inline_response_200_2"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
    post:
      description: "This operation is used for creating a new subscription to zone\
        \ status change notification."
      operationId: "subscriptions_zonal_status_post"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "zoneStatusSubscription"
        description: "Zone Status Subscription"
        required: true
        schema:
          $ref: "#/definitions/ZoneStatusSubscription"
      responses:
        "201":
          description: "Response to create new zone status subscription"
          examples:
            application/json:
              zoneStatusSubscription:
                clientCorrelator: "0123"
                resourceURL: "http://example.com/exampleAPI/location/v1/subscriptions/zoneStatus/subscription123"
                callbackReference:
                  notifyURL: "http://clientApp.example.com/location_notifications/123456"
                zoneId: "zone01"
                numberOfUsersZoneThreshold: "500"
                operationStatus: "Serviceable"
          schema:
            properties:
              zonalTrafficSubscription:
                $ref: "#/definitions/ZoneStatusSubscription"
      x-swagger-router-controller: "swagger_server.controllers.default_controller"
...
definitions:
...
  AccessPointInfo:
    type: "object"
    required:
    - "accessPointId"
    - "connectionType"
    - "numberOfUsers"
    - "operationStatus"
    - "resourceURL"
    properties:
      accessPointId:
        $ref: "#/definitions/AccessPointId"
      locationInfo:
        $ref: "#/definitions/LocationInfo"
      connectionType:
        $ref: "#/definitions/ConnectionType"
      operationStatus:
        $ref: "#/definitions/OperationStatus"
      numberOfUsers:
        $ref: "#/definitions/NumberOfUsers"
      timezone:
        $ref: "#/definitions/Timezone"
      interestRealm:
        $ref: "#/definitions/InterestRealm"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
    description: "A type containing access point information."
  AccessPointList:
    type: "object"
    required:
    - "resourceURL"
    - "zoneId"
    properties:
      zoneId:
        $ref: "#/definitions/ZoneId"
      accessPoint:
        type: "array"
        description: "Collection of the access point information list."
        items:
          $ref: "#/definitions/AccessPointInfo"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
    description: "A type containing list of access points."
...
  ZoneInfo:
    type: "object"
    required:
    - "numberOfAccessPoints"
    - "numberOfUnservicableAccessPoints"
    - "numberOfUsers"
    - "resourceURL"
    - "zoneId"
    properties:
      zoneId:
        $ref: "#/definitions/ZoneId"
      numberOfAccessPoints:
        $ref: "#/definitions/NumberOfAccessPoints"
      numberOfUnservicableAccessPoints:
        $ref: "#/definitions/NumberOfUnserviceableAccessPoints"
      numberOfUsers:
        $ref: "#/definitions/NumberOfUsers"
      resourceURL:
        $ref: "#/definitions/ResourceURL"
    description: "A type containing zone information."
...
  inline_response_200:
    type: "object"
    properties:
      notificationSubscriptionList:
        $ref: "#/definitions/inline_response_200_notificationSubscriptionList"
...